refactor: simplify component props by extending Action types in Explainer, Final, Info, Playlist, Score, and Trial components
